% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidarLDA_foldin.R
\name{LidarLDA_foldin}
\alias{LidarLDA_foldin}
\title{Title}
\usage{
LidarLDA_foldin(
  y,
  n,
  nclust,
  gamma,
  ngibbs,
  nburn,
  phi.post,
  phi.estim,
  theta.post
)
}
\arguments{
\item{y}{P x H matrix containing the number of returns for each pixel in each height class}

\item{n}{P x H matrix containing the number of incoming light pulses for each pixel in each height class}

\item{nclust}{maximum number of clusters (K)}

\item{gamma}{parameter between 0 and 1 for the prior of the truncated stick-breaking prior}

\item{ngibbs}{number of iterations for the MCMC algorithm}

\item{nburn}{number of iterations to discard as burn in}

\item{phi.post}{logical value (T or F) indicating if samples from the posterior distribution for phi are being provided or not}

\item{phi.estim}{if phi.post is true, then the algorithm
expects that this will be a large matrix containing
samples from the posterior distribution for phi. Each row of
this matrix should contain a separate sample. If phi.post is false,
then this function expects that this is a K x H matrix
containing an estimate (e.g., the posterior mean) of phi.}

\item{theta.post}{should samples from the posterior distribution for theta be returned (TRUE or FALSE)? If FALSE, just the posterior mean is returned}
}
\value{
This function returns a list containing several elements:
\itemize{
\item llk:  log-likelihood for each iteration. This is a vector of size ngibbs.
\item theta: estimated relative abundance of each cluster in each pixel.
This consists of a matrix where rows are samples
from the posterior distribution.
}
}
\description{
Title
}
